move browse to separate module, make `metadata` group
